+++ crates/rethnet_eth/src/remote/client.rs
@@ -6,7 +6,14 @@ use std::{
 use itertools::Itertools;
 use revm_primitives::{AccountInfo, Address, Bytecode, B256, KECCAK_EMPTY, U256};
 
-use super::{eth, jsonrpc, BlockSpec, GetLogsInput, MethodInvocation, ZeroXPrefixedBytes};
+use super::{
+    eth, jsonrpc,
+    methods::{
+        eth::{EthMethodInvocation, GetLogsInput},
+        MethodInvocation,
+    },
+    BlockSpec, ZeroXPrefixedBytes,
+};
 
 /// Specialized error types
 #[derive(thiserror::Error, Debug)]
@@ -206,9 +213,12 @@ impl RpcClient {
         block: BlockSpec,
     ) -> Result<AccountInfo, RpcClientError> {
         let inputs = Vec::from([
-            MethodInvocation::GetBalance(*address, block.clone()),
-            MethodInvocation::GetTransactionCount(*address, block.clone()),
-            MethodInvocation::GetCode(*address, block),
+            MethodInvocation::Eth(EthMethodInvocation::GetBalance(*address, block.clone())),
+            MethodInvocation::Eth(EthMethodInvocation::GetTransactionCount(
+                *address,
+                block.clone(),
+            )),
+            MethodInvocation::Eth(EthMethodInvocation::GetCode(*address, block)),
         ]);
 
         let response = self.batch_call(&inputs).await?;
@@ -333,8 +343,10 @@ impl RpcClient {
         &self,
         hash: &B256,
     ) -> Result<Option<eth::Block<B256>>, RpcClientError> {
-        self.call(&MethodInvocation::GetBlockByHash(*hash, false))
-            .await
+        self.call(&MethodInvocation::Eth(EthMethodInvocation::GetBlockByHash(
+            *hash, false,
+        )))
+        .await
     }
 
     /// Calls `eth_getBlockByHash` and returns the transaction's data.
@@ -342,8 +354,10 @@ impl RpcClient {
         &self,
         hash: &B256,
     ) -> Result<Option<eth::Block<eth::Transaction>>, RpcClientError> {
-        self.call(&MethodInvocation::GetBlockByHash(*hash, true))
-            .await
+        self.call(&MethodInvocation::Eth(EthMethodInvocation::GetBlockByHash(
+            *hash, true,
+        )))
+        .await
     }
 
     /// Calls `eth_getBlockByNumber` and returns the transaction's hash.
@@ -351,8 +365,10 @@ impl RpcClient {
         &self,
         spec: BlockSpec,
     ) -> Result<eth::Block<B256>, RpcClientError> {
-        self.call(&MethodInvocation::GetBlockByNumber(spec, false))
-            .await
+        self.call(&MethodInvocation::Eth(
+            EthMethodInvocation::GetBlockByNumber(spec, false),
+        ))
+        .await
     }
 
     /// Calls `eth_getBlockByNumber` and returns the transaction's data.
@@ -360,8 +376,10 @@ impl RpcClient {
         &self,
         spec: BlockSpec,
     ) -> Result<eth::Block<eth::Transaction>, RpcClientError> {
-        self.call(&MethodInvocation::GetBlockByNumber(spec, true))
-            .await
+        self.call(&MethodInvocation::Eth(
+            EthMethodInvocation::GetBlockByNumber(spec, true),
+        ))
+        .await
     }
 
     /// eth_getLogs
@@ -371,11 +389,13 @@ impl RpcClient {
         to_block: BlockSpec,
         address: &Address,
     ) -> Result<Vec<eth::Log>, RpcClientError> {
-        self.call(&MethodInvocation::GetLogs(GetLogsInput {
-            from_block,
-            to_block,
-            address: *address,
-        }))
+        self.call(&MethodInvocation::Eth(EthMethodInvocation::GetLogs(
+            GetLogsInput {
+                from_block,
+                to_block,
+                address: *address,
+            },
+        )))
         .await
     }
 
@@ -384,8 +404,10 @@ impl RpcClient {
         &self,
         tx_hash: &B256,
     ) -> Result<Option<eth::Transaction>, RpcClientError> {
-        self.call(&MethodInvocation::GetTransactionByHash(*tx_hash))
-            .await
+        self.call(&MethodInvocation::Eth(
+            EthMethodInvocation::GetTransactionByHash(*tx_hash),
+        ))
+        .await
     }
 
     /// eth_getTransactionCount
@@ -394,8 +416,10 @@ impl RpcClient {
         address: &Address,
         block: BlockSpec,
     ) -> Result<U256, RpcClientError> {
-        self.call(&MethodInvocation::GetTransactionCount(*address, block))
-            .await
+        self.call(&MethodInvocation::Eth(
+            EthMethodInvocation::GetTransactionCount(*address, block),
+        ))
+        .await
     }
 
     /// eth_getTransactionReceipt
@@ -403,8 +427,10 @@ impl RpcClient {
         &self,
         tx_hash: &B256,
     ) -> Result<Option<eth::TransactionReceipt>, RpcClientError> {
-        self.call(&MethodInvocation::GetTransactionReceipt(*tx_hash))
-            .await
+        self.call(&MethodInvocation::Eth(
+            EthMethodInvocation::GetTransactionReceipt(*tx_hash),
+        ))
+        .await
     }
 
     /// eth_getStorageAt
@@ -414,8 +440,10 @@ impl RpcClient {
         position: U256,
         block: BlockSpec,
     ) -> Result<U256, RpcClientError> {
-        self.call(&MethodInvocation::GetStorageAt(*address, position, block))
-            .await
+        self.call(&MethodInvocation::Eth(EthMethodInvocation::GetStorageAt(
+            *address, position, block,
+        )))
+        .await
     }
 }
 
@@ -445,7 +473,9 @@ mod tests {
                 .expect("failed to parse hash from string");
 
         let error = RpcClient::new(&server.url())
-            .call::<Option<eth::Transaction>>(&MethodInvocation::GetTransactionByHash(hash))
+            .call::<Option<eth::Transaction>>(&MethodInvocation::Eth(
+                EthMethodInvocation::GetTransactionByHash(hash),
+            ))
             .await
             .expect_err("should have failed to interpret response as a Transaction");
 
@@ -488,7 +518,9 @@ mod tests {
             .expect("failed to parse hash from string");
 
             let error = RpcClient::new(alchemy_url)
-                .call::<Option<eth::Transaction>>(&MethodInvocation::GetTransactionByHash(hash))
+                .call::<Option<eth::Transaction>>(&MethodInvocation::Eth(
+                    EthMethodInvocation::GetTransactionByHash(hash),
+                ))
                 .await
                 .expect_err("should have failed to interpret response as a Transaction");
 
@@ -509,7 +541,9 @@ mod tests {
             .expect("failed to parse hash from string");
 
             let error = RpcClient::new(alchemy_url)
-                .call::<Option<eth::Transaction>>(&MethodInvocation::GetTransactionByHash(hash))
+                .call::<Option<eth::Transaction>>(&MethodInvocation::Eth(
+                    EthMethodInvocation::GetTransactionByHash(hash),
+                ))
                 .await
                 .expect_err("should have failed to connect due to a garbage domain name");
