+++ crates/rethnet_eth/src/remote/methods.rs
@@ -1,4 +1,5 @@
-use crate::remote::MethodInvocation as EthMethodInvocation;
+/// Ethereum-standard RPC methods
+pub mod eth;
 
 /// RPC methods specific to Hardhat
 pub mod hardhat;
@@ -9,7 +10,7 @@ pub mod hardhat;
 #[allow(clippy::large_enum_variant)]
 pub enum MethodInvocation {
     /// an eth_* method invocation
-    Eth(EthMethodInvocation),
+    Eth(eth::EthMethodInvocation),
     /// a hardhat_* method invocation
     Hardhat(hardhat::HardhatMethodInvocation),
 }
